Refactor AppPage to utilize Wizard component, simplifying the structure and enhancing the user experience. Clean up imports and streamline the page rendering process.
